refactor: extract event routes into ui/events.py
Move 16 event route endpoints and 3 helper functions from the 4890-line
routes.py into a dedicated ui/events.py module (1140 lines removed from
routes.py). Convert routes to use FastAPI dependency injection:

- Manual auth checks → Depends(require_user) from ui/common
- _get_db context manager → Depends(get_db) from dependencies
- Manual context building → base_context() from ui/common
- HTMX request detection → render_fragment/trigger_event from ui/htmx
- _PAGE_SIZE → PAGE_SIZE from ui/common

Also:
- Fix trigger_event to use Python 3.12+ type parameter syntax (UP047)
- Move fastapi.templating out of TYPE_CHECKING block (TC004)
- Add TC003 suppression for ui/*.py (FastAPI resolves Annotated at runtime)
- Update test_artist_enrichment to include events router

Session 2 of 5 for the component system refactor.
