+++ src/resonance/ui/routes.py
@@ -5,7 +5,7 @@ import pathlib
 import uuid
 import zoneinfo
 from contextlib import asynccontextmanager
-from typing import TYPE_CHECKING, Annotated, Any
+from typing import TYPE_CHECKING
 
 import fastapi
 import fastapi.requests
@@ -16,10 +16,6 @@ import sqlalchemy as sa
 import sqlalchemy.ext.asyncio as sa_async
 import sqlalchemy.orm as sa_orm
 
-import resonance.concerts.sync as concert_sync
-import resonance.connectors.listenbrainz as listenbrainz_module
-import resonance.connectors.spotify as spotify_module
-import resonance.crypto as crypto_module
 import resonance.dedup as dedup_module
 import resonance.dependencies as deps_module
 import resonance.merge as merge_module
@@ -31,7 +27,6 @@ import resonance.models.playlist as playlist_models
 import resonance.models.task as task_models
 import resonance.models.taste as taste_models
 import resonance.models.user as user_models
-import resonance.services.artist_utils as artist_utils
 import resonance.types as types_module
 import resonance.ui.filters as filters_module
 import resonance.ui.view_filters as view_filters_module
@@ -781,1180 +776,6 @@ async def artist_merge_preview(
             "events_to_repoint": events_to_repoint,
             "merged_links": merged_links,
         },
-    )
-
-
-@router.get("/events", response_model=None)
-async def events_page(
-    request: fastapi.Request,
-    page: int = 1,
-) -> fastapi.responses.HTMLResponse | fastapi.responses.RedirectResponse:
-    """Render paginated events list with venue and artist info, or redirect to login."""
-    user_id = request.state.session.get("user_id")
-    if not user_id:
-        return fastapi.responses.RedirectResponse(url="/login", status_code=307)
-
-    user_uuid = uuid.UUID(user_id)
-    offset = (page - 1) * _PAGE_SIZE
-
-    # Parse filter parameters
-    params = dict(request.query_params)
-    multi_params = {
-        "attendance": request.query_params.getlist("attendance"),
-        "source_service": request.query_params.getlist("source_service"),
-    }
-
-    # Build presets (today's date resolved dynamically)
-    presets = view_filters_module.build_event_presets()
-    _event_filter_keys = {
-        "q",
-        "title",
-        "venue",
-        "artist",
-        "date_from",
-        "date_to",
-        "attendance",
-        "source_service",
-        "has_pending",
-        "include_not_going",
-    }
-    active_preset = view_filters_module.detect_active_preset(
-        params,
-        presets,
-        filter_keys=_event_filter_keys,
-        default_preset="upcoming",
-    )
-
-    # If "upcoming" is the default preset and no date_from was explicitly set,
-    # inject today's date so the filter is applied.
-    if active_preset == "upcoming" and "date_from" not in params:
-        params["date_from"] = view_filters_module._today_iso()
-
-    # Parse filters to get active_filters for template context
-    applied = filters_module.parse_filter_params(
-        view_filters_module.EVENT_FILTERS, params, multi_params=multi_params
-    )
-
-    async with _get_db(request) as db:
-        # Build base query with joins for cross-entity filtering
-        query = (
-            sa.select(concert_models.Event)
-            .outerjoin(
-                concert_models.Venue,
-                concert_models.Event.venue_id == concert_models.Venue.id,
-            )
-            .outerjoin(
-                concert_models.EventArtist,
-                concert_models.Event.id == concert_models.EventArtist.event_id,
-            )
-            .outerjoin(
-                music_models.Artist,
-                concert_models.EventArtist.artist_id == music_models.Artist.id,
-            )
-        )
-
-        # Apply registered filter fields (title, venue, artist, date, has_pending)
-        query = filters_module.apply_filters(
-            query,
-            view_filters_module.EVENT_FILTERS,
-            params,
-            multi_params=multi_params,
-        )
-
-        # Handle attendance filter manually (requires user_id context)
-        attendance_values = multi_params.get("attendance", [])
-        _valid_set = {"GOING", "INTERESTED", "NOT_GOING", "UNSET"}
-        valid_attendance = [v for v in attendance_values if v in _valid_set]
-        if valid_attendance:
-            status_values = [v for v in valid_attendance if v != "UNSET"]
-            include_unset = "UNSET" in valid_attendance
-            conditions: list[sa.ColumnElement[bool]] = []
-            if status_values:
-                attendance_subquery = sa.select(
-                    concert_models.UserEventAttendance.event_id
-                ).where(
-                    concert_models.UserEventAttendance.user_id == user_uuid,
-                    concert_models.UserEventAttendance.status.in_(status_values),
-                )
-                conditions.append(concert_models.Event.id.in_(attendance_subquery))
-            if include_unset:
-                has_attendance = sa.select(
-                    concert_models.UserEventAttendance.event_id
-                ).where(
-                    concert_models.UserEventAttendance.user_id == user_uuid,
-                )
-                conditions.append(concert_models.Event.id.not_in(has_attendance))
-            query = query.where(sa.or_(*conditions))
-        elif params.get("include_not_going") != "true":
-            # Default: exclude NOT_GOING events
-            not_going_subquery = sa.select(
-                concert_models.UserEventAttendance.event_id
-            ).where(
-                concert_models.UserEventAttendance.user_id == user_uuid,
-                concert_models.UserEventAttendance.status == "NOT_GOING",
-            )
-            query = query.where(concert_models.Event.id.not_in(not_going_subquery))
-
-        # Deduplicate rows from outer joins, order, and paginate
-        query = (
-            query.group_by(concert_models.Event.id)
-            .order_by(concert_models.Event.event_date.desc())
-            .offset(offset)
-            .limit(_PAGE_SIZE + 1)
-        )
-
-        # Wrap in a subquery so eager loads work correctly with group_by
-        event_ids_subquery = query.with_only_columns(concert_models.Event.id)
-        final_query = (
-            sa.select(concert_models.Event)
-            .where(concert_models.Event.id.in_(event_ids_subquery))
-            .options(
-                sa_orm.joinedload(concert_models.Event.venue),
-                sa_orm.joinedload(concert_models.Event.artists),
-                sa_orm.joinedload(concert_models.Event.artist_candidates),
-            )
-            .order_by(concert_models.Event.event_date.desc())
-        )
-
-        result = await db.execute(final_query)
-        events = list(result.unique().scalars().all())
-
-        has_next = len(events) > _PAGE_SIZE
-        events = events[:_PAGE_SIZE]
-
-        event_ids = [e.id for e in events]
-        attendance_map = await _get_attendance_map(db, user_uuid, event_ids)
-
-    # Build filter query string for pagination links
-    filter_qs = view_filters_module.build_filter_query_string(
-        applied.active_filters, view_filters_module.EVENT_FILTERS
-    )
-    # Include attendance in filter_qs (handled outside the standard fields)
-    if valid_attendance:
-        attendance_parts = [f"attendance={v}" for v in valid_attendance]
-        if filter_qs:
-            filter_qs += "&" + "&".join(attendance_parts)
-        else:
-            filter_qs = "&".join(attendance_parts)
-    # Include quick search in filter_qs
-    q_value = params.get("q", "").strip()
-    if q_value:
-        if filter_qs:
-            filter_qs += f"&q={q_value}"
-        else:
-            filter_qs = f"q={q_value}"
-
-    # Build flat active_filters dict for the template
-    template_active_filters: dict[str, object] = {}
-    for key, value in applied.active_filters.items():
-        if isinstance(value, dict):
-            # DateRangeField stores {date_from: ..., date_to: ...}
-            for dk, dv in value.items():
-                if dv is not None:
-                    template_active_filters[dk] = str(dv)
-        else:
-            template_active_filters[key] = value
-    if valid_attendance:
-        template_active_filters["attendance"] = valid_attendance
-    if q_value:
-        template_active_filters["q"] = q_value
-
-    context = {
-        "user_id": user_id,
-        "user_tz": _user_tz(request),
-        "user_role": _user_role(request),
-        "events": events,
-        "attendance_map": attendance_map,
-        "page": page,
-        "has_next": has_next,
-        "has_prev": page > 1,
-        "active_filters": template_active_filters,
-        "presets": presets,
-        "filters": view_filters_module.EVENT_TEMPLATE_FILTERS,
-        "active_preset": active_preset,
-        "list_url": "/events",
-        "list_target": "#event-list",
-        "filter_qs": filter_qs,
-    }
-
-    if request.headers.get("HX-Request"):
-        return templates.TemplateResponse(request, "partials/event_list.html", context)
-    return templates.TemplateResponse(request, "events.html", context)
-
-
-@router.get("/events/{event_id}", response_model=None)
-async def event_detail_page(
-    request: fastapi.Request,
-    event_id: uuid.UUID,
-) -> fastapi.responses.HTMLResponse | fastapi.responses.RedirectResponse:
-    """Render event detail page with artists, candidates, and add-artist search."""
-    user_id = request.state.session.get("user_id")
-    if not user_id:
-        return fastapi.responses.RedirectResponse(url="/login", status_code=307)
-
-    async with _get_db(request) as db:
-        event_result = await db.execute(
-            sa.select(concert_models.Event)
-            .where(concert_models.Event.id == event_id)
-            .options(
-                sa_orm.joinedload(concert_models.Event.venue),
-                sa_orm.joinedload(concert_models.Event.artists),
-                sa_orm.joinedload(concert_models.Event.artist_candidates),
-            )
-        )
-        event = event_result.unique().scalar_one_or_none()
-
-        if event is None:
-            raise fastapi.HTTPException(status_code=404, detail="Event not found")
-
-        # Re-match any unmatched candidates against the current artist catalog
-        has_unmatched = any(
-            c.matched_artist_id is None
-            for c in event.artist_candidates
-            if c.status == types_module.CandidateStatus.PENDING
-        )
-        if has_unmatched:
-            matched = await concert_sync.match_candidates_to_artists(db, event)
-            if matched > 0:
-                await db.commit()
-                await db.refresh(event, ["artist_candidates"])
-
-        # Build dict of matched_artist_id -> Artist for candidates with matches
-        matched_ids = [
-            c.matched_artist_id
-            for c in event.artist_candidates
-            if c.matched_artist_id is not None
-        ]
-        matched_artists: dict[uuid.UUID, music_models.Artist] = {}
-        if matched_ids:
-            artists_result = await db.execute(
-                sa.select(music_models.Artist).where(
-                    music_models.Artist.id.in_(matched_ids)
-                )
-            )
-            for a in artists_result.scalars().all():
-                matched_artists[a.id] = a
-
-        attendance = (
-            await db.execute(
-                sa.select(concert_models.UserEventAttendance).where(
-                    concert_models.UserEventAttendance.user_id == uuid.UUID(user_id),
-                    concert_models.UserEventAttendance.event_id == event_id,
-                )
-            )
-        ).scalar_one_or_none()
-
-    context = {
-        "user_id": user_id,
-        "user_tz": _user_tz(request),
-        "user_role": _user_role(request),
-        "event": event,
-        "matched_artists": matched_artists,
-        "attendance": attendance,
-    }
-
-    return templates.TemplateResponse(request, "event_detail.html", context)
-
-
-def _normalize_positions(
-    artists: list[concert_models.EventArtist],
-) -> list[concert_models.EventArtist]:
-    """Assign sequential positions (0, 1, 2, ...) based on current sort order."""
-    sorted_artists = sorted(artists, key=lambda ea: ea.position)
-    for i, ea in enumerate(sorted_artists):
-        ea.position = i
-    return sorted_artists
-
-
-@router.post("/events/{event_id}/artists/{ea_id}/move-up", response_model=None)
-async def move_artist_up(
-    request: fastapi.Request,
-    event_id: uuid.UUID,
-    ea_id: uuid.UUID,
-) -> fastapi.responses.HTMLResponse | fastapi.responses.RedirectResponse:
-    """Swap an event artist with the one above it and return updated list."""
-    user_id = request.state.session.get("user_id")
-    if not user_id:
-        return fastapi.responses.RedirectResponse(url="/login", status_code=307)
-
-    async with _get_db(request) as db:
-        event = await _load_event_with_artists(db, event_id)
-        sorted_artists = _normalize_positions(list(event.artists))
-        target_idx = next(
-            (i for i, ea in enumerate(sorted_artists) if ea.id == ea_id), None
-        )
-        if target_idx is not None and target_idx > 0:
-            sorted_artists[target_idx - 1].position = target_idx
-            sorted_artists[target_idx].position = target_idx - 1
-        await db.commit()
-
-    return templates.TemplateResponse(
-        request,
-        "partials/event_confirmed_artists.html",
-        {"event": event},
-    )
-
-
-@router.post("/events/{event_id}/artists/{ea_id}/move-down", response_model=None)
-async def move_artist_down(
-    request: fastapi.Request,
-    event_id: uuid.UUID,
-    ea_id: uuid.UUID,
-) -> fastapi.responses.HTMLResponse | fastapi.responses.RedirectResponse:
-    """Swap an event artist with the one below it and return updated list."""
-    user_id = request.state.session.get("user_id")
-    if not user_id:
-        return fastapi.responses.RedirectResponse(url="/login", status_code=307)
-
-    async with _get_db(request) as db:
-        event = await _load_event_with_artists(db, event_id)
-        sorted_artists = _normalize_positions(list(event.artists))
-        target_idx = next(
-            (i for i, ea in enumerate(sorted_artists) if ea.id == ea_id), None
-        )
-        if target_idx is not None and target_idx < len(sorted_artists) - 1:
-            sorted_artists[target_idx].position = target_idx + 1
-            sorted_artists[target_idx + 1].position = target_idx
-        await db.commit()
-
-    return templates.TemplateResponse(
-        request,
-        "partials/event_confirmed_artists.html",
-        {"event": event},
-    )
-
-
-@router.post("/events/{event_id}/artists/{ea_id}/remove", response_model=None)
-async def remove_artist_from_event(
-    request: fastapi.Request,
-    event_id: uuid.UUID,
-    ea_id: uuid.UUID,
-) -> fastapi.responses.HTMLResponse | fastapi.responses.RedirectResponse:
-    """Remove a confirmed artist from an event and return updated list."""
-    user_id = request.state.session.get("user_id")
-    if not user_id:
-        return fastapi.responses.RedirectResponse(url="/login", status_code=307)
-
-    async with _get_db(request) as db:
-        ea = (
-            await db.execute(
-                sa.select(concert_models.EventArtist).where(
-                    concert_models.EventArtist.id == ea_id,
-                    concert_models.EventArtist.event_id == event_id,
-                )
-            )
-        ).scalar_one_or_none()
-        if ea is not None:
-            await db.delete(ea)
-            await db.commit()
-
-        event = await _load_event_with_artists(db, event_id)
-        _normalize_positions(list(event.artists))
-        await db.commit()
-
-    return templates.TemplateResponse(
-        request,
-        "partials/event_confirmed_artists.html",
-        {"event": event},
-    )
-
-
-@router.get("/events/{event_id}/artists", response_model=None)
-async def event_artists_partial(
-    request: fastapi.Request,
-    event_id: uuid.UUID,
-) -> fastapi.responses.HTMLResponse | fastapi.responses.RedirectResponse:
-    """Return confirmed artists partial for HTMX refresh."""
-    user_id = request.state.session.get("user_id")
-    if not user_id:
-        return fastapi.responses.RedirectResponse(url="/login", status_code=307)
-
-    async with _get_db(request) as db:
-        event = await _load_event_with_artists(db, event_id)
-
-    return templates.TemplateResponse(
-        request,
-        "partials/event_confirmed_artists.html",
-        {"event": event},
-    )
-
-
-async def _load_event_with_artists(
-    db: sa_async.AsyncSession,
-    event_id: uuid.UUID,
-) -> concert_models.Event:
-    result = await db.execute(
-        sa.select(concert_models.Event)
-        .where(concert_models.Event.id == event_id)
-        .options(sa_orm.joinedload(concert_models.Event.artists))
-    )
-    event = result.unique().scalar_one_or_none()
-    if event is None:
-        raise fastapi.HTTPException(status_code=404, detail="Event not found")
-    return event
-
-
-async def _get_attendance_map(
-    db: sa_async.AsyncSession,
-    user_id: uuid.UUID,
-    event_ids: list[uuid.UUID],
-) -> dict[uuid.UUID, concert_models.UserEventAttendance]:
-    if not event_ids:
-        return {}
-    result = await db.execute(
-        sa.select(concert_models.UserEventAttendance).where(
-            concert_models.UserEventAttendance.user_id == user_id,
-            concert_models.UserEventAttendance.event_id.in_(event_ids),
-        )
-    )
-    return {a.event_id: a for a in result.scalars().all()}
-
-
-@router.post("/events/{event_id}/attendance", response_model=None)
-async def set_attendance(
-    request: fastapi.Request,
-    event_id: uuid.UUID,
-    status: Annotated[str, fastapi.Form()],
-) -> fastapi.responses.HTMLResponse | fastapi.responses.RedirectResponse:
-    """Set user's attendance status for an event, returns updated partial."""
-    user_id = request.state.session.get("user_id")
-    if not user_id:
-        return fastapi.responses.RedirectResponse(url="/login", status_code=307)
-
-    user_uuid = uuid.UUID(user_id)
-    new_status = types_module.AttendanceStatus(status)
-
-    async with _get_db(request) as db:
-        existing = (
-            await db.execute(
-                sa.select(concert_models.UserEventAttendance).where(
-                    concert_models.UserEventAttendance.user_id == user_uuid,
-                    concert_models.UserEventAttendance.event_id == event_id,
-                )
-            )
-        ).scalar_one_or_none()
-
-        if existing is not None:
-            existing.status = new_status
-            existing.source_service = types_module.ServiceType.MANUAL
-            attendance = existing
-        else:
-            attendance = concert_models.UserEventAttendance(
-                user_id=user_uuid,
-                event_id=event_id,
-                status=new_status,
-                source_service=types_module.ServiceType.MANUAL,
-            )
-            db.add(attendance)
-
-        await db.commit()
-
-    return templates.TemplateResponse(
-        request,
-        "partials/attendance_status.html",
-        {"attendance": attendance, "event_id": event_id},
-    )
-
-
-@router.get("/events/{event_id}/candidates", response_model=None)
-async def event_candidates_partial(
-    request: fastapi.Request,
-    event_id: uuid.UUID,
-) -> fastapi.responses.HTMLResponse | fastapi.responses.RedirectResponse:
-    """Return candidates partial for HTMX refresh."""
-    user_id = request.state.session.get("user_id")
-    if not user_id:
-        return fastapi.responses.RedirectResponse(url="/login", status_code=307)
-
-    async with _get_db(request) as db:
-        event = (
-            (
-                await db.execute(
-                    sa.select(concert_models.Event)
-                    .where(concert_models.Event.id == event_id)
-                    .options(
-                        sa_orm.joinedload(concert_models.Event.artist_candidates),
-                    )
-                )
-            )
-            .unique()
-            .scalar_one()
-        )
-        matched_artist_ids = [
-            c.matched_artist_id
-            for c in event.artist_candidates
-            if c.matched_artist_id is not None
-        ]
-        matched_artists: dict[uuid.UUID, music_models.Artist] = {}
-        if matched_artist_ids:
-            ma_result = await db.execute(
-                sa.select(music_models.Artist).where(
-                    music_models.Artist.id.in_(matched_artist_ids)
-                )
-            )
-            for a in ma_result.scalars().all():
-                matched_artists[a.id] = a
-
-    return templates.TemplateResponse(
-        request,
-        "partials/event_candidates.html",
-        {"event": event, "matched_artists": matched_artists},
-    )
-
-
-@router.post("/events/{event_id}/candidates/{candidate_id}/accept", response_model=None)
-async def accept_candidate_ui(
-    request: fastapi.Request,
-    event_id: uuid.UUID,
-    candidate_id: uuid.UUID,
-) -> fastapi.responses.HTMLResponse | fastapi.responses.RedirectResponse:
-    """Accept an artist candidate and create a confirmed EventArtist."""
-    user_id = request.state.session.get("user_id")
-    if not user_id:
-        return fastapi.responses.RedirectResponse(url="/login", status_code=307)
-
-    async with _get_db(request) as db:
-        candidate = (
-            await db.execute(
-                sa.select(concert_models.EventArtistCandidate).where(
-                    concert_models.EventArtistCandidate.id == candidate_id,
-                    concert_models.EventArtistCandidate.event_id == event_id,
-                )
-            )
-        ).scalar_one_or_none()
-
-        if candidate is None:
-            raise fastapi.HTTPException(status_code=404, detail="Candidate not found")
-
-        if candidate.matched_artist_id is None:
-            raise fastapi.HTTPException(
-                status_code=400, detail="Candidate has no matched artist"
-            )
-
-        event_artist = concert_models.EventArtist(
-            event_id=event_id,
-            artist_id=candidate.matched_artist_id,
-            position=candidate.position,
-            raw_name=candidate.raw_name,
-        )
-        db.add(event_artist)
-        candidate.status = types_module.CandidateStatus.ACCEPTED
-        await db.commit()
-
-    response = templates.TemplateResponse(
-        request, "partials/candidate_accepted.html", {"candidate": candidate}
-    )
-    response.headers["HX-Trigger"] = "artistsChanged"
-    return response
-
-
-@router.post("/events/{event_id}/candidates/{candidate_id}/reject", response_model=None)
-async def reject_candidate_ui(
-    request: fastapi.Request,
-    event_id: uuid.UUID,
-    candidate_id: uuid.UUID,
-) -> fastapi.responses.HTMLResponse | fastapi.responses.RedirectResponse:
-    """Reject an artist candidate."""
-    user_id = request.state.session.get("user_id")
-    if not user_id:
-        return fastapi.responses.RedirectResponse(url="/login", status_code=307)
-
-    async with _get_db(request) as db:
-        candidate = (
-            await db.execute(
-                sa.select(concert_models.EventArtistCandidate).where(
-                    concert_models.EventArtistCandidate.id == candidate_id,
-                    concert_models.EventArtistCandidate.event_id == event_id,
-                )
-            )
-        ).scalar_one_or_none()
-
-        if candidate is None:
-            raise fastapi.HTTPException(status_code=404, detail="Candidate not found")
-
-        candidate.status = types_module.CandidateStatus.REJECTED
-        await db.commit()
-
-    return templates.TemplateResponse(
-        request, "partials/candidate_rejected.html", {"candidate": candidate}
-    )
-
-
-@router.post("/events/{event_id}/add-artist", response_model=None)
-async def add_artist_to_event_ui(
-    request: fastapi.Request,
-    event_id: uuid.UUID,
-    artist_id: Annotated[uuid.UUID, fastapi.Form()],
-    candidate_id: Annotated[uuid.UUID | None, fastapi.Form()] = None,
-) -> fastapi.responses.HTMLResponse:
-    """Create a candidate from artist search and return feedback partial."""
-    user_id = request.state.session.get("user_id")
-    if not user_id:
-        return fastapi.responses.HTMLResponse("Unauthorized", status_code=401)
-
-    async with _get_db(request) as db:
-        artist = (
-            await db.execute(
-                sa.select(music_models.Artist).where(
-                    music_models.Artist.id == artist_id
-                )
-            )
-        ).scalar_one_or_none()
-        if artist is None:
-            return fastapi.responses.HTMLResponse(
-                '<small style="color: var(--pico-del-color);">Artist not found</small>'
-            )
-
-        # Check if already confirmed as EventArtist
-        already_confirmed = (
-            await db.execute(
-                sa.select(concert_models.EventArtist).where(
-                    concert_models.EventArtist.event_id == event_id,
-                    concert_models.EventArtist.artist_id == artist_id,
-                )
-            )
-        ).scalar_one_or_none()
-        # Resolve the specific candidate if provided
-        if candidate_id is not None:
-            target_candidate = (
-                await db.execute(
-                    sa.select(concert_models.EventArtistCandidate).where(
-                        concert_models.EventArtistCandidate.id == candidate_id,
-                        concert_models.EventArtistCandidate.event_id == event_id,
-                    )
-                )
-            ).scalar_one_or_none()
-            if target_candidate is not None:
-                target_candidate.matched_artist_id = artist.id
-                target_candidate.confidence_score = 100
-                target_candidate.status = types_module.CandidateStatus.ACCEPTED
-
-        if already_confirmed is not None:
-            if candidate_id is not None and target_candidate is not None:
-                await db.commit()
-            # Reload and return updated candidates
-            event = (
-                (
-                    await db.execute(
-                        sa.select(concert_models.Event)
-                        .where(concert_models.Event.id == event_id)
-                        .options(
-                            sa_orm.joinedload(concert_models.Event.artist_candidates),
-                        )
-                    )
-                )
-                .unique()
-                .scalar_one()
-            )
-            matched_artist_ids = [
-                c.matched_artist_id
-                for c in event.artist_candidates
-                if c.matched_artist_id is not None
-            ]
-            early_matched: dict[uuid.UUID, music_models.Artist] = {}
-            if matched_artist_ids:
-                ma_result = await db.execute(
-                    sa.select(music_models.Artist).where(
-                        music_models.Artist.id.in_(matched_artist_ids)
-                    )
-                )
-                for a in ma_result.scalars().all():
-                    early_matched[a.id] = a
-            resp = templates.TemplateResponse(
-                request,
-                "partials/event_candidates.html",
-                {"event": event, "matched_artists": early_matched},
-            )
-            resp.headers["HX-Trigger"] = "artistsChanged"
-            return resp
-
-        # Check for any existing candidate with this raw_name
-        existing = (
-            await db.execute(
-                sa.select(concert_models.EventArtistCandidate).where(
-                    concert_models.EventArtistCandidate.event_id == event_id,
-                    concert_models.EventArtistCandidate.raw_name == artist.name,
-                )
-            )
-        ).scalar_one_or_none()
-
-        if existing is not None:
-            existing.matched_artist_id = artist.id
-            existing.confidence_score = 100
-            existing.status = types_module.CandidateStatus.ACCEPTED
-            candidate = existing
-        else:
-            candidate = concert_models.EventArtistCandidate(
-                event_id=event_id,
-                raw_name=artist.name,
-                matched_artist_id=artist.id,
-                status=types_module.CandidateStatus.ACCEPTED,
-                confidence_score=100,
-            )
-            db.add(candidate)
-
-        event_artist = concert_models.EventArtist(
-            event_id=event_id,
-            artist_id=artist.id,
-            position=candidate.position or 0,
-            raw_name=artist.name,
-        )
-        db.add(event_artist)
-        await db.commit()
-
-        # Reload event with candidates for refreshing the candidates section
-        event = (
-            (
-                await db.execute(
-                    sa.select(concert_models.Event)
-                    .where(concert_models.Event.id == event_id)
-                    .options(
-                        sa_orm.joinedload(concert_models.Event.artist_candidates),
-                    )
-                )
-            )
-            .unique()
-            .scalar_one()
-        )
-
-        matched_artist_ids = [
-            c.matched_artist_id
-            for c in event.artist_candidates
-            if c.matched_artist_id is not None
-        ]
-        matched_artists: dict[uuid.UUID, music_models.Artist] = {}
-        if matched_artist_ids:
-            ma_result = await db.execute(
-                sa.select(music_models.Artist).where(
-                    music_models.Artist.id.in_(matched_artist_ids)
-                )
-            )
-            for a in ma_result.scalars().all():
-                matched_artists[a.id] = a
-
-    response = templates.TemplateResponse(
-        request,
-        "partials/event_candidates.html",
-        {"event": event, "matched_artists": matched_artists},
-    )
-    response.headers["HX-Trigger"] = "artistsChanged"
-    return response
-
-
-@router.get("/partials/artist-search", response_model=None)
-async def artist_search_partial(
-    request: fastapi.Request,
-    q: str = "",
-    event_id: uuid.UUID | None = None,
-) -> fastapi.responses.HTMLResponse | fastapi.responses.RedirectResponse:
-    """Search artists by name and return results partial for HTMX."""
-    user_id = request.state.session.get("user_id")
-    if not user_id:
-        return fastapi.responses.RedirectResponse(url="/login", status_code=307)
-
-    if len(q.strip()) < 2:
-        return fastapi.responses.HTMLResponse("")
-
-    async with _get_db(request) as db:
-        result = await db.execute(
-            sa.select(music_models.Artist)
-            .where(music_models.Artist.name.ilike(f"%{_escape_ilike(q.strip())}%"))
-            .order_by(music_models.Artist.name)
-            .limit(10)
-        )
-        artists = list(result.scalars().all())
-
-    return templates.TemplateResponse(
-        request,
-        "partials/artist_search_results.html",
-        {
-            "artists": artists,
-            "event_id": event_id,
-            "query": q.strip(),
-        },
-    )
-
-
-@router.get("/partials/artist-search-modal", response_model=None)
-async def artist_search_modal_partial(
-    request: fastapi.Request,
-    q: str = "",
-    event_id: uuid.UUID | None = None,
-    candidate_id: uuid.UUID | None = None,
-) -> fastapi.responses.HTMLResponse | fastapi.responses.RedirectResponse:
-    """Serve the external search modal dialog."""
-    user_id = request.state.session.get("user_id")
-    if not user_id:
-        return fastapi.responses.RedirectResponse(url="/login", status_code=307)
-
-    connected_services: list[str] = []
-    async with _get_db(request) as db:
-        stmt = sa.select(user_models.ServiceConnection.service_type).where(
-            user_models.ServiceConnection.user_id == uuid.UUID(user_id),
-            user_models.ServiceConnection.service_type
-            == types_module.ServiceType.SPOTIFY,
-        )
-        result = await db.execute(stmt)
-        if result.scalar_one_or_none() is not None:
-            connected_services.append("spotify")
-
-    return templates.TemplateResponse(
-        request,
-        "partials/artist_search_modal.html",
-        {
-            "query": q,
-            "event_id": event_id,
-            "candidate_id": candidate_id,
-            "connected_services": connected_services,
-        },
-    )
-
-
-@router.get("/partials/artist-search-external", response_model=None)
-async def artist_search_external_partial(
-    request: fastapi.Request,
-    q: str = "",
-    event_id: uuid.UUID | None = None,
-    candidate_id: uuid.UUID | None = None,
-) -> fastapi.responses.HTMLResponse | fastapi.responses.RedirectResponse:
-    """Search external services and return results partial for HTMX."""
-    user_id = request.state.session.get("user_id")
-    if not user_id:
-        return fastapi.responses.RedirectResponse(url="/login", status_code=307)
-
-    registry = request.app.state.connector_registry
-    lb = registry.get_base_connector(types_module.ServiceType.LISTENBRAINZ)
-    query = q.strip()
-
-    results: list[dict[str, Any]] = []
-    detected_service: str | None = None
-
-    # Check if query is a recognized service URL
-    url_connector = None
-    url_id: str | None = None
-    if query.startswith(("http://", "https://")):
-        for connector in registry.all_base_connectors():
-            parsed = connector.parse_url(query)
-            if parsed is not None:
-                url_connector = connector
-                url_id = parsed
-                break
-
-    if url_connector is not None and url_id is not None:
-        detected_service = url_connector.service_type.value
-        if isinstance(url_connector, listenbrainz_module.ListenBrainzConnector):
-            detected_service = "musicbrainz"
-            mb_artist = await url_connector.get_artist_by_mbid(url_id)
-            if mb_artist:
-                results = [mb_artist]
-        elif isinstance(url_connector, spotify_module.SpotifyConnector):
-            async with _get_db(request) as db:
-                conn_result = await db.execute(
-                    sa.select(user_models.ServiceConnection).where(
-                        user_models.ServiceConnection.user_id == uuid.UUID(user_id),
-                        user_models.ServiceConnection.service_type
-                        == types_module.ServiceType.SPOTIFY,
-                    )
-                )
-                spotify_conn = conn_result.scalar_one_or_none()
-                if spotify_conn and spotify_conn.encrypted_access_token:
-                    settings = request.app.state.settings
-                    token = crypto_module.decrypt_token(
-                        spotify_conn.encrypted_access_token,
-                        settings.token_encryption_key,
-                    )
-                    if (
-                        spotify_conn.token_expires_at is not None
-                        and spotify_conn.token_expires_at
-                        <= datetime.datetime.now(datetime.UTC)
-                        and spotify_conn.encrypted_refresh_token is not None
-                    ):
-                        refresh = crypto_module.decrypt_token(
-                            spotify_conn.encrypted_refresh_token,
-                            settings.token_encryption_key,
-                        )
-                        tok_resp = await url_connector.refresh_access_token(refresh)
-                        token = tok_resp.access_token
-                        spotify_conn.encrypted_access_token = (
-                            crypto_module.encrypt_token(
-                                token, settings.token_encryption_key
-                            )
-                        )
-                        if tok_resp.expires_in is not None:
-                            spotify_conn.token_expires_at = datetime.datetime.now(
-                                datetime.UTC
-                            ) + datetime.timedelta(seconds=tok_resp.expires_in)
-                        await db.commit()
-                    artist_data = await url_connector.get_artist_by_id(token, url_id)
-                    if artist_data:
-                        results.append(
-                            {
-                                "mbid": "",
-                                "name": artist_data["name"],
-                                "disambiguation": "",
-                                "artist_type": "",
-                                "area": "",
-                                "begin_year": None,
-                                "end_year": None,
-                                "source": "spotify",
-                                "spotify_id": artist_data["spotify_id"],
-                            }
-                        )
-    elif len(query) < 2:
-        return fastapi.responses.HTMLResponse("")
-    elif lb:
-        mb_results = await lb.search_artists(query, limit=10)
-        results = mb_results
-        async with _get_db(request) as db:
-            for r in mb_results:
-                stmt = sa.select(music_models.Artist).where(
-                    sa.or_(
-                        music_models.Artist.service_links["musicbrainz"][
-                            "id"
-                        ].as_string()
-                        == r["mbid"],
-                        music_models.Artist.service_links["listenbrainz"].as_string()
-                        == r["mbid"],
-                    )
-                )
-                result = await db.execute(stmt)
-                existing = result.scalar_one_or_none()
-                r["already_imported"] = existing is not None
-                r["local_artist_id"] = str(existing.id) if existing else None
-                r["already_on_event"] = False
-                if existing is not None and event_id is not None:
-                    on_event = await db.execute(
-                        sa.select(concert_models.EventArtist.id).where(
-                            concert_models.EventArtist.event_id == event_id,
-                            concert_models.EventArtist.artist_id == existing.id,
-                        )
-                    )
-                    r["already_on_event"] = on_event.scalar_one_or_none() is not None
-            results = mb_results
-
-    return templates.TemplateResponse(
-        request,
-        "partials/artist_external_results.html",
-        {
-            "external_artists": results,
-            "event_id": event_id,
-            "candidate_id": candidate_id,
-            "detected_service": detected_service,
-        },
-    )
-
-
-@router.post("/partials/artist-import", response_model=None)
-async def artist_import_partial(
-    request: fastapi.Request,
-    mbid: Annotated[str, fastapi.Form()] = "",
-    spotify_id: Annotated[str, fastapi.Form()] = "",
-    name: Annotated[str, fastapi.Form()] = "",
-    disambiguation: Annotated[str, fastapi.Form()] = "",
-    artist_type: Annotated[str, fastapi.Form()] = "",
-    area: Annotated[str, fastapi.Form()] = "",
-    begin_year: Annotated[str, fastapi.Form()] = "",
-    end_year: Annotated[str, fastapi.Form()] = "",
-    event_id: Annotated[uuid.UUID | None, fastapi.Form()] = None,
-    candidate_id: Annotated[uuid.UUID | None, fastapi.Form()] = None,
-) -> fastapi.responses.HTMLResponse | fastapi.responses.RedirectResponse:
-    """Import an artist from external search and return a local result row."""
-    user_id = request.state.session.get("user_id")
-    if not user_id:
-        return fastapi.responses.RedirectResponse(url="/login", status_code=307)
-
-    async with _get_db(request) as db:
-        artist: music_models.Artist | None = None
-
-        # Check for existing artist by MBID or Spotify ID
-        dedup_conditions: list[sa.ColumnElement[bool]] = []
-        if mbid:
-            dedup_conditions.append(
-                music_models.Artist.service_links["musicbrainz"]["id"].as_string()
-                == mbid
-            )
-            dedup_conditions.append(
-                music_models.Artist.service_links["listenbrainz"].as_string() == mbid
-            )
-        if spotify_id:
-            dedup_conditions.append(
-                music_models.Artist.service_links["spotify"]["id"].as_string()
-                == spotify_id
-            )
-        if dedup_conditions:
-            result = await db.execute(
-                sa.select(music_models.Artist).where(sa.or_(*dedup_conditions))
-            )
-            artist = result.scalar_one_or_none()
-
-        if artist is None:
-            links: dict[str, Any] = {}
-            if mbid:
-                links["musicbrainz"] = {"id": mbid}
-            if spotify_id:
-                links["spotify"] = {"id": spotify_id}
-            artist = music_models.Artist(
-                name=name,
-                disambiguation=disambiguation or None,
-                artist_type=artist_type or None,
-                area=area or None,
-                begin_year=int(begin_year) if begin_year else None,
-                end_year=int(end_year) if end_year else None,
-                service_links=links,
-            )
-            db.add(artist)
-            await db.flush()
-
-        if event_id:
-            # Check if already on event
-            already = (
-                await db.execute(
-                    sa.select(concert_models.EventArtist.id).where(
-                        concert_models.EventArtist.event_id == event_id,
-                        concert_models.EventArtist.artist_id == artist.id,
-                    )
-                )
-            ).scalar_one_or_none()
-
-            if already is None:
-                # Resolve specific candidate if provided
-                if candidate_id:
-                    candidate = (
-                        await db.execute(
-                            sa.select(concert_models.EventArtistCandidate).where(
-                                concert_models.EventArtistCandidate.id == candidate_id,
-                                concert_models.EventArtistCandidate.event_id
-                                == event_id,
-                            )
-                        )
-                    ).scalar_one_or_none()
-                    if candidate is not None:
-                        candidate.matched_artist_id = artist.id
-                        candidate.confidence_score = 100
-                        candidate.status = types_module.CandidateStatus.ACCEPTED
-
-                # Also check for unresolved candidate by name
-                if not candidate_id:
-                    existing_candidate = (
-                        await db.execute(
-                            sa.select(concert_models.EventArtistCandidate).where(
-                                concert_models.EventArtistCandidate.event_id
-                                == event_id,
-                                concert_models.EventArtistCandidate.raw_name
-                                == artist.name,
-                                concert_models.EventArtistCandidate.status
-                                == types_module.CandidateStatus.PENDING,
-                            )
-                        )
-                    ).scalar_one_or_none()
-                    if existing_candidate is not None:
-                        existing_candidate.matched_artist_id = artist.id
-                        existing_candidate.confidence_score = 100
-                        existing_candidate.status = (
-                            types_module.CandidateStatus.ACCEPTED
-                        )
-
-                db.add(
-                    concert_models.EventArtist(
-                        event_id=event_id,
-                        artist_id=artist.id,
-                        position=0,
-                        raw_name=artist.name,
-                    )
-                )
-
-        await db.commit()
-
-    response = fastapi.responses.HTMLResponse("")
-    response.headers["HX-Trigger"] = '{"artist-imported":"", "artistsChanged":""}'
-    return response
-
-
-_ENRICHMENT_STALENESS_SECONDS = 180
-
-
-@router.get("/partials/artist-enrich/{artist_id}", response_model=None)
-async def artist_enrich_partial(
-    request: fastapi.Request,
-    artist_id: uuid.UUID,
-    event_id: uuid.UUID | None = None,
-) -> fastapi.responses.HTMLResponse | fastapi.responses.RedirectResponse:
-    """Lazily enrich an artist with MusicBrainz metadata."""
-    user_id = request.state.session.get("user_id")
-    if not user_id:
-        return fastapi.responses.RedirectResponse(url="/login", status_code=307)
-
-    async with _get_db(request) as db:
-        result = await db.execute(
-            sa.select(music_models.Artist).where(music_models.Artist.id == artist_id)
-        )
-        artist = result.scalar_one_or_none()
-        if artist is None:
-            return fastapi.responses.HTMLResponse("")
-
-        mbid = artist_utils.get_mbid(artist.service_links)
-
-        # Already enriched or no MBID — return as-is
-        if not mbid or artist.disambiguation is not None:
-            return templates.TemplateResponse(
-                request,
-                "partials/artist_row.html",
-                {"artist": artist, "event_id": event_id},
-            )
-
-        # Check enrichment_requested_at timestamp
-        mb_data = (artist.service_links or {}).get("musicbrainz", {})
-        requested_at_str = (
-            mb_data.get("enrichment_requested_at")
-            if isinstance(mb_data, dict)
-            else None
-        )
-        if requested_at_str:
-            requested_at = datetime.datetime.fromisoformat(requested_at_str)
-            elapsed = (
-                datetime.datetime.now(datetime.UTC) - requested_at
-            ).total_seconds()
-            if elapsed < _ENRICHMENT_STALENESS_SECONDS:
-                # Recent request, skip
-                return templates.TemplateResponse(
-                    request,
-                    "partials/artist_row.html",
-                    {"artist": artist, "event_id": event_id},
-                )
-
-        # Mark enrichment requested
-        links = dict(artist.service_links or {})
-        mb: dict[str, Any] = (
-            dict(links.get("musicbrainz", {}))
-            if isinstance(links.get("musicbrainz"), dict)
-            else {}
-        )
-        mb["enrichment_requested_at"] = datetime.datetime.now(datetime.UTC).isoformat()
-        if mbid and "id" not in mb:
-            mb["id"] = mbid
-        links["musicbrainz"] = mb
-        artist.service_links = links
-        await db.commit()
-
-        # Fetch from MusicBrainz
-        registry = request.app.state.connector_registry
-        lb = registry.get_base_connector(types_module.ServiceType.LISTENBRAINZ)
-        if lb:
-            mb_artist = await lb.get_artist_by_mbid(mbid)
-            if mb_artist:
-                artist.disambiguation = mb_artist.get("disambiguation") or ""
-                artist.artist_type = mb_artist.get("artist_type") or None
-                artist.area = mb_artist.get("area") or None
-                artist.begin_year = mb_artist.get("begin_year")
-                artist.end_year = mb_artist.get("end_year")
-            else:
-                artist.disambiguation = ""
-            await db.commit()
-
-    return templates.TemplateResponse(
-        request,
-        "partials/artist_row.html",
-        {"artist": artist, "event_id": event_id},
     )
